requirements.txt
streamlit==1.35.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pandas==2.2.2
openpyxl==3.1.2
streamlit-aggrid==0.3.4.post3
python-dotenv==1.0.1
